requirements.txt
Django==3.1
django-bootstrap4==2.2.0
django-model-utils==4.0.0
django-crispy-forms==1.9.2